ECOCKPIT-62 plat: imx8: new conversion from mpidr to core index
PSCI use table indexed by each CPU cores.

On i.MX8QM, the table is indexed this way:
[idx0: A53_0][idx1: A53_1][idx2: A53_2][idx3: A53_3][idx4: A72_0][idx5: A72_1]

For eCockpit, an instance of ATF has only access to a cluster.
Therefore, depending on which cluster ATF runs, the table is either indexed
this way:
[idx0: A53_0][idx 1: A53_1][idx2: A53_2][idx3: A53_3]
or this way:
[idx0: A72_0][idx1: A72_1]

This patch rewrites functions plat_core_pos_by_mpidr() and
plat_my_core_pos to match this indexing strategy.

Signed-off-by: Fabrice Goucem <fabrice.goucem@nxp.com>